feat(editor): add pixel-level vertical scrolling
Support immediate trackpad scrolling and half-row stops. Keep rendering,
hit testing, IME, popups, and scrollbars aligned with the viewport.

Handle caret visibility, scroll bounds, and layout changes consistently.
Add regression coverage for input, scrolling, and fractional display
scales.

+++ app/src/editor/scintillaimageitem_p.h
@@ -21,7 +21,7 @@ class ScintillaImageItem : public QQuickItem
     explicit ScintillaImageItem(QQuickItem* parent = nullptr);
     void invalidateImage(const QRectF& rect);
     void invalidateImage();
-    void scrollImage(int dy);
+    void scrollImage(qreal dy);
 
   protected:
     virtual bool paintImage(QPainter& painter, const QRect& rect) = 0;
